chore: moved cookie CRUD helpers out of /cookie route
